Refactor database service methods to use SELECT instead of CALL for stored procedures
- Updated BookingDbService methods to use SELECT for invoking stored procedures.
- Changed CatalogDbService methods to utilize SELECT for stored procedure calls.
- Modified InventoryDbService methods to replace CALL with SELECT for stored procedures.
- Adjusted PaymentDbService to use SELECT instead of CALL for updating payment status.
- Updated PayrollDbService methods to invoke stored procedures using SELECT.
- Refactored ProfileDbService methods to use SELECT for updating user profiles.
- Changed ReviewDbService to correctly pass customerId instead of createdBy in review creation.
- Updated SchedulingDbService methods to use SELECT for scheduling operations.
- Refactored StaffDbService to use DateOnly for workDate parameter and updated attendance logging.
- Enhanced SqlExceptionExtensions to correctly identify application-level errors based on SQLSTATE.
- Modified PostgreSQL stored procedures to ensure correct handling of parameters and return types.

+++ backend/SaloonApi/Modules/Reports/Infrastructure/ReportsRepository.cs
@@ -3,11 +3,15 @@ using SaloonApi.Shared.Data.DbServices;
 
 namespace SaloonApi.Modules.Reports.Infrastructure;
 
-internal sealed record SalesByServiceDto(int TreatmentId, string TreatmentName, string CategoryName, int BookingCount, decimal TotalRevenue);
-internal sealed record SalesByStaffDto(int TherapistId, string TherapistName, int BookingCount, decimal TotalRevenue);
-internal sealed record SalesByLocationDto(int LocationId, string LocationName, int BookingCount, decimal TotalRevenue);
-internal sealed record RetentionDto(int TotalCustomers, int ReturningCustomers, decimal RetentionRatePercent);
-internal sealed record NoShowRateDto(int TotalAppointments, int NoShowCount, decimal NoShowRatePercent);
+// BookingCount/TotalCustomers/etc are COUNT(*)/SUM(int) results -- Postgres returns bigint for
+// those, which Npgsql/Dapper materializes as long, not int (positional-record materialization
+// requires an exact type match, so `int` here throws "no matching constructor" the moment a row
+// actually comes back).
+internal sealed record SalesByServiceDto(int TreatmentId, string TreatmentName, string CategoryName, long BookingCount, decimal TotalRevenue);
+internal sealed record SalesByStaffDto(int TherapistId, string TherapistName, long BookingCount, decimal TotalRevenue);
+internal sealed record SalesByLocationDto(int LocationId, string LocationName, long BookingCount, decimal TotalRevenue);
+internal sealed record RetentionDto(long TotalCustomers, long ReturningCustomers, decimal RetentionRatePercent);
+internal sealed record NoShowRateDto(long TotalAppointments, long NoShowCount, decimal NoShowRatePercent);
 
 internal sealed class ReportsRepository(SqlConnectionFactory factory, ReportsDbService reportsDb)
 {
